Starting to implement multicontroller setup
I have started to implement multicontroller setup. It is nearly complete but there are a couple of steps that I have to take. Known bugs are the Rift S controllers are named the same as the CV1 controllers so the CV1 controllers show up. There is only the right hand at the moment.